AIML-539: Move reconcile_open_remediations to domains/workflow/
Extract from main.py into src/smartfix/domains/workflow/pr_reconciliation.py
to follow the existing pattern where workflow steps are defined in the
workflow domain and called from the entry point.
